Prod and dev env (#3)
The Dockerfile was modified in order to support multistages.
One stage is for the dev version and the other one for the env.

The docker-compose.yml file is used for the prod version, while docker-compose.dev.yml is used for the development one. The latter one detects changes in the code and reruns the app.